Separate speed vizualizer -> waypoint_marker/waypoint_velocity_vizualizer
