+++ ros/src/computing/planning/motion/packages/waypoint_maker/launch/waypoint_loader.launch
@@ -14,8 +14,6 @@
         <param name="disableVelocitySmoothing" value="$(arg disableVelocitySmoothing)" />
 	</node>
 
-	<node pkg="waypoint_maker" type="waypoint_marker_publisher" name="waypoint_marker_publisher">
-	<param name="use_velocity_visualizer" value="$(arg use_velocity_visualizer)"/>
-	</node>
+	<node pkg="waypoint_maker" type="waypoint_marker_publisher" name="waypoint_marker_publisher" />
 
 </launch>
