+++ ros/src/util/packages/runtime_manager/scripts/computing.yaml
@@ -594,6 +594,17 @@ subs :
          desc : waypoint_clicker desc sample
          cmd  : rosrun waypoint_maker waypoint_clicker
          param: waypoint_clicker
+       - name : waypoint_velocity_vizualizer
+         desc : waypoint_velocity_vizualizer desc sample
+         cmd  : roslaunch waypoint_maker waypoint_velocity_vizualizer.launch
+         param: waypoint_velocity_vizualizer
+         gui  :
+           lane_waypoint_csv :
+             prop  : 1
+           decelerate :
+                   flags : [nl ]
+           disableDecisionMaker :
+                   flags : [nl ]
     - name : waypoint_follower
       desc : waypoint_follower desc sample
       subs :
@@ -2942,15 +2953,6 @@ params :
       label   : 'decelerate :'
       v         : 1.0
       cmd_param :
-        dash        : ''
-        delim       : ':='
-    - name      : use_velocity_visualizer
-      desc      : velocity visualizer desc sample
-      label     : Velocity Visualizer
-      kind      : checkbox
-      v         : False
-      cmd_param :
-        var_name    : use_velocity_visualizer
         dash        : ''
         delim       : ':='
     - name      : disableDecisionMaker
